move runtime.spec
