+++ packages/com/test/tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "extends": "../../../tsconfig.test.json",
+  "compilerOptions": {
+    "outDir": "../dist/test"
+  },
+  "references": [
+    {
+      "path": "../src"
+    }
+  ]
+}
